+++ printf/sources/ft_puthex_pf.c
@@ -6,15 +6,15 @@
 /*   By: kvan-don <kvan-don@student.42luxembourg.l  +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/03/14 14:37:56 by kvan-don          #+#    #+#             */
-/*   Updated: 2024/03/14 16:07:38 by kvan-don         ###   ########.fr       */
+/*   Updated: 2024/03/15 15:44:10 by kvan-don         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
-#include "ft_printf.h"
+#include "../ft_printf.h"
 
-void	ft_puthex_pf(unsigned int num, size_t *count, char base)
+void	ft_puthex_pf(unsigned long num, size_t *count, char base)
 {
-	if (num >= 16) 
+	if (num >= 16)
 	{
 		ft_puthex_pf(num / 16, count, base);
 		ft_puthex_pf(num % 16, count, base);
@@ -30,5 +30,5 @@ void	ft_puthex_pf(unsigned int num, size_t *count, char base)
 			else if (base == 'X')
 				ft_putchar_pf(num - 10 + 'A', count);
 		}
-	}	
+	}
 }
